requirements.txt
torch
numpy
scipy
matplotlib
pydantic
fastapi
uvicorn
onnx
onnxruntime
boto3
botocore
arm_pyart
